move asset search test to integration tests
